De aca voy a tratar de agregarle las categorias
