+++ src/apps/blog/forms.py
@@ -8,7 +8,16 @@ class Formulario_Alta_Post(forms.ModelForm):
 
     class Meta:
         model = Post
-        fields = '__all__'
+        fields = ('titulo', 'contenido', 'id_user', 'categoria')
+
+        widgets = {
+        	'titulo': forms.TextInput(attrs={'class': 'form-control'}),
+        	'id_user': forms.TextInput(attrs={'class': 'form-control', 'value':'', 'id':'identificador', 'type':'hidden'}),
+        	'contenido': forms.Textarea(attrs={'class': 'form-control'}),
+           	'categoria': forms.Select(attrs={'class': 'form-control'}),
+
+
+        }
 
 class Formulario_Alta_Comentario(forms.ModelForm):
 	class Meta:
@@ -17,8 +26,8 @@ class Formulario_Alta_Comentario(forms.ModelForm):
 
 
 
-class Formulario_Registro_Usuario(UserCreationForm):
-    class Meta:
-        model = Usuario
-        fields = ['username', 'email', 'first_name', 'last_name', 'password1', 'password2']
+#class Formulario_Registro_Usuario(UserCreationForm):
+ #   class Meta:
+  #      model = Usuario
+   #     fields = ['username', 'email', 'first_name', 'last_name', 'password1', 'password2']
