+++ src/apps/blog/models.py
@@ -1,12 +1,20 @@
 from django.db import models
 from django.utils import timezone
 from datetime import datetime
+from django.contrib.auth.models import AbstractUser
 
 
 # Creacion de modelos
+
+#class Usuario(AbstractUser):
+ #   pass
+    #id=models.AutoField(primary_key=True)
+    # nacionalidad =models.CharField(max_length=30)
+
+
 class Post(models.Model):
     id=models.AutoField(primary_key=True)
-    #id_user = models.ForeignKey('Usuario',on_delete=models.CASCADE)
+    id_user = models.ForeignKey('usuario.Usuario',on_delete=models.CASCADE)
     titulo = models.CharField(max_length=100)
     contenido = models.TextField()
     fecha_publicacion = models.DateTimeField(auto_now_add=True)
