+++ src/apps/blog/urls.py
@@ -12,6 +12,6 @@ urlpatterns = [
   #path('nuevo/', post_nuevo, name = 'nuevo'),
   path('editar/<int:pk>/',views.Editar_post.as_view(), name= 'editar'),
   path('eliminar/<int:pk>/',eliminar_post, name= 'eliminar'),
-  path('crear_usuario/', views.Alta_usuario.as_view(), name='crear_usuario')
+  #path('crear_usuario/', views.Alta_usuario.as_view(), name='crear_usuario')
  
 ]
