+++ src/apps/blog/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect
 from django.views.generic import *
 from django.urls import reverse_lazy, reverse
 from .models import Post, Comentario
-from .forms import Formulario_Alta_Post, Formulario_Alta_Comentario, Formulario_Registro_Usuario
+from .forms import Formulario_Alta_Post, Formulario_Alta_Comentario
 from django.http import HttpResponse, HttpResponseRedirect
 from ..usuario.models import Usuario
 
@@ -102,8 +102,8 @@ class Alta_post(LoginRequiredMixin ,CreateView):
 	# 		form.save()
 	# 		return redirect('home')
 
-class Alta_usuario(CreateView):
-    model = Usuario
-    form_class = Formulario_Registro_Usuario
-    template_name = 'usuario/altaUsuario.html'
-    success_url = reverse_lazy('home')
+#class Alta_usuario(CreateView):
+#    model = Usuario
+ #   form_class = Formulario_Registro_Usuario
+  #  template_name = 'usuario/altaUsuario.html'
+   # #success_url = reverse_lazy('home')
