+++ src/tpfinal/urls.py
@@ -23,8 +23,10 @@ urlpatterns = [
     path('', include('apps.blog.urls')),
     # path('posts', include('apps.blog.urls')),
     path('admin/', admin.site.urls),
-    path('about/', views.about, name ='about'),
-    path('categoria/', views.categoria, name='categorias'),
+    path('about/', views.about, name ='about'),    
+    path('usuario', include('django.contrib.auth.urls')),
+    path('', include('apps.usuario.urls')),
+    #path('categoria/', views.categoria, name='categorias'),
 
     path('login/',auth.LoginView.as_view(template_name="usuario/login.html"), name='login'),
     path('logout/',auth.LogoutView.as_view(), name="logout"),
